requirements.txt
fastapi==0.115.8
joblib==1.4.2
mlflow==2.20.2
mlflow_skinny==2.20.2
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
pytest==8.3.4
PyYAML==6.0.2
s3fs==2025.2.0
scikit_learn==1.6.1
scipy==1.15.2
xgboost==2.1.4
